requirements.txt
django 
djangorestframework 
django-cors-headers
pillow
bs4
plotly
django-allauth
tzdata

pydantic==1.10.9
blosc2==2.0.0
cython==0.29.21
black==22.3.0
langchain
langchain_experimental
chromadb
openai
playwright
pypdf
finance-datareader
tabulate